+++ CHANGELOG.md
@@ -17,9 +17,11 @@ This version includes all revisions from r90 to rXX.
 - OutpostHD will now create a 'savegame' directory under 'data' if one doesn't exist.
 - Minimap now displays location of active robots.
 - User can now change the tube selection when placing tubes using the wheel mouse (allows for quick tube placement).
+- Using the mouse's scroll wheel when in tube insertion mode will now switch which tube piece is being placed.
 
 ### Fixed
 - Fixed a mistake in the way robots were handled in the UI when deployed. Robot selection UI will now properly update itself when any robot in the field completes its task instead of the original robots created when the SEED Lander is deployed.
+- Fixed a dangling pointer bug when deserializing robots.